Add files via upload
I changed the base images in the assets folder. These images were the app icon images for the browser, expo site project file, and android app icon etc. I replaced those images the app hagakure icon but with the original png names and sizes. So where as previously, the image icon would refer to the expo icon, now it refers to the hagkure icon. Where as previously the favicon image was some sort of cube (which appears when you open the app via the browser), now favicon image refers to the hagakure icon but it has the same name 'favicon' and same pixel sizes as the previous image. For future applications, follow this procedure of replacing the 4 images in the assets folder, with your 'icon' image while utilizing that image-name and image-size that was previously provided. To generate the same 'icon' image in different sizes, i used this site: https://derivv.com/.  Don't forget, the image-names and image-sizes cannot change, the only thing changing is the actual image.